refactor: extract useNavigation hook
